+++ cmd/neofs-lens/internal/meta/root.go
@@ -34,6 +34,7 @@ func init() {
 		listGraveyardCMD,
 		listGarbageCMD,
 		writeObjectCMD,
+		getCMD,
 	)
 }