+++ cmd/neofs-lens/internal/storage/root.go
@@ -44,7 +44,7 @@ var Root = &cobra.Command{
 
 func init() {
 	Root.AddCommand(
-		storageInspectObjCMD,
+		storageInspectObjCMD, storageGetObjCMD,
 	)
 }